fix: address PR review findings for basic auth
- Fix btoa stack overflow: replace spread-into-fromCharCode with reduce
  in createBasicAuth and updateBasicAuth; safe for any htpasswd size
- Fix useUpdateHttpProxy options generic to carry { previous } context
  type so onError callback is correctly typed end-to-end
- Show "Enabled" instead of "0 users" in config card list view where
  the htpasswd Secret is not fetched (intentional trade-off to avoid N+1)
- Move basicAuthSchema and BasicAuthSchema type to http-proxy.schema.ts
  so validation rules are co-located with other domain schemas and
  reusable/testable in isolation; export via index
- Document SHA-1 constraint in basicAuthSchema comment (only scheme
  supported by Envoy Gateway's BasicAuth filter)
