+++ app/resources/http-proxies/http-proxy.queries.ts
@@ -57,7 +57,7 @@ export function useCreateHttpProxy(
 export function useUpdateHttpProxy(
   projectId: string,
   name: string,
-  options?: UseMutationOptions<HttpProxy, Error, UpdateHttpProxyInput>
+  options?: UseMutationOptions<HttpProxy, Error, UpdateHttpProxyInput, { previous: HttpProxy | undefined }>
 ) {
   const queryClient = useQueryClient();
